Removes weapon abilities from the core
