+++ Projects/Server.Tests/Tests/Network/Packets/Outgoing/PlayerPacketTests.cs
@@ -57,22 +57,6 @@ namespace Server.Tests.Network
         }
 
         [Theory]
-        [InlineData(0, true)]
-        [InlineData(0, false)]
-        [InlineData(100, true)]
-        [InlineData(1000, false)]
-        public void TestSpecialAbility(int abilityId, bool active)
-        {
-            var expected = new ToggleSpecialAbility(abilityId, active).Compile();
-
-            var ns = PacketTestUtilities.CreateTestNetState();
-            ns.SendToggleSpecialAbility(abilityId, active);
-
-            var result = ns.SendPipe.Reader.TryRead();
-            AssertThat.Equal(result.Buffer[0].AsSpan(0), expected);
-        }
-
-        [Theory]
         [InlineData(0x1000u, "This is a header", "This is a body", "This is a footer")]
         [InlineData(0x1000u, null, null, null)]
         public void TestDisplayProfile(uint serial, string header, string body, string footer)
@@ -333,18 +317,6 @@ namespace Server.Tests.Network
 
             var ns = PacketTestUtilities.CreateTestNetState();
             ns.SendPingAck(ping);
-
-            var result = ns.SendPipe.Reader.TryRead();
-            AssertThat.Equal(result.Buffer[0].AsSpan(0), expected);
-        }
-
-        [Fact]
-        public void TestClearAbility()
-        {
-            var expected = new ClearWeaponAbility().Compile();
-
-            var ns = PacketTestUtilities.CreateTestNetState();
-            ns.SendClearWeaponAbility();
 
             var result = ns.SendPipe.Reader.TryRead();
             AssertThat.Equal(result.Buffer[0].AsSpan(0), expected);
