+++ Projects/Server.Tests/Tests/Network/Packets/Outgoing/PlayerPackets.cs
@@ -57,19 +57,6 @@ namespace Server.Network
         public static Packet Instantiate(bool dead) => dead ? Dead : Alive;
     }
 
-    public sealed class ToggleSpecialAbility : Packet
-    {
-        public ToggleSpecialAbility(int abilityID, bool active) : base(0xBF)
-        {
-            EnsureCapacity(7);
-
-            Stream.Write((short)0x25);
-
-            Stream.Write((short)abilityID);
-            Stream.Write(active);
-        }
-    }
-
     public sealed class DisplayProfile : Packet
     {
         public DisplayProfile(Serial m, string header, string body, string footer) : base(0xB8)
@@ -371,18 +358,6 @@ namespace Server.Network
             }
 
             return p;
-        }
-    }
-
-    public sealed class ClearWeaponAbility : Packet
-    {
-        public static readonly Packet Instance = SetStatic(new ClearWeaponAbility());
-
-        public ClearWeaponAbility() : base(0xBF)
-        {
-            EnsureCapacity(5);
-
-            Stream.Write((short)0x21);
         }
     }
