+++ Projects/UOContent/Items/Weapons/Abilities/WeaponAbility.cs
@@ -132,15 +132,17 @@ namespace Server.Items
         {
             var mana = BaseMana;
 
-            var skillTotal = GetSkill(from, SkillName.Swords) + GetSkill(from, SkillName.Macing)
-                                                              + GetSkill(from, SkillName.Fencing) +
-                                                              GetSkill(from, SkillName.Archery) +
-                                                              GetSkill(from, SkillName.Parry)
-                                                              + GetSkill(from, SkillName.Lumberjacking) +
-                                                              GetSkill(from, SkillName.Stealth)
-                                                              + GetSkill(from, SkillName.Poisoning) +
-                                                              GetSkill(from, SkillName.Bushido) +
-                                                              GetSkill(from, SkillName.Ninjitsu);
+            var skillTotal =
+                GetSkill(from, SkillName.Swords) +
+                GetSkill(from, SkillName.Macing) +
+                GetSkill(from, SkillName.Fencing) +
+                GetSkill(from, SkillName.Archery) +
+                GetSkill(from, SkillName.Parry) +
+                GetSkill(from, SkillName.Lumberjacking) +
+                GetSkill(from, SkillName.Stealth) +
+                GetSkill(from, SkillName.Poisoning) +
+                GetSkill(from, SkillName.Bushido) +
+                GetSkill(from, SkillName.Ninjitsu);
 
             if (skillTotal >= 300.0)
             {
@@ -196,9 +198,11 @@ namespace Server.Items
             }
 
             /* <UBWS> */
-            if (weapon.WeaponAttributes.UseBestSkill > 0 && (from.Skills.Swords.Base >= reqSkill ||
-                                                             from.Skills.Macing.Base >= reqSkill ||
-                                                             from.Skills.Fencing.Base >= reqSkill))
+            if (weapon.WeaponAttributes.UseBestSkill > 0 &&
+                (from.Skills.Swords.Base >= reqSkill ||
+                 from.Skills.Macing.Base >= reqSkill ||
+                 from.Skills.Fencing.Base >= reqSkill)
+            )
             {
                 return true;
             }
@@ -394,23 +398,6 @@ namespace Server.Items
             }
         }
 
-        public static void Initialize()
-        {
-            EventSink.SetAbility += EventSink_SetAbility;
-        }
-
-        private static void EventSink_SetAbility(Mobile m, int index)
-        {
-            if (index == 0)
-            {
-                ClearCurrentAbility(m);
-            }
-            else if (index >= 1 && index < Abilities.Length)
-            {
-                SetCurrentAbility(m, Abilities[index]);
-            }
-        }
-
         private static void AddContext(Mobile m, WeaponAbilityContext context)
         {
             m_PlayersTable[m] = context;
@@ -418,19 +405,10 @@ namespace Server.Items
 
         private static void RemoveContext(Mobile m)
         {
-            var context = GetContext(m);
-
-            if (context != null)
+            if (m_PlayersTable.Remove(m, out var context))
             {
-                RemoveContext(m, context);
+                context.Timer?.Stop();
             }
-        }
-
-        private static void RemoveContext(Mobile m, WeaponAbilityContext context)
-        {
-            m_PlayersTable.Remove(m);
-
-            context.Timer.Stop();
         }
 
         private static WeaponAbilityContext GetContext(Mobile m)
