+++ Projects/UOContent/Spells/Bushido/SamuraiSpell.cs
@@ -1,6 +1,6 @@
 using System;
+using Server.Items;
 using Server.Mobiles;
-using Server.Network;
 
 namespace Server.Spells.Bushido
 {
